Reformat the Settings UI
