+++ src/app/desktop/utils/middlewareInstances.js
@@ -41,10 +41,10 @@ const middleware = (store) => (next) => (action) => {
         // eslint-disable-next-line
         notification.open({
           key: "nsfwNotWorking",
-          message: "NSFW Error",
+          message: "Error Code 13",
           description: "Node Sentinel File Watcher could not be initialized",
           top: 47,
-          duration: 10,
+          duration: 45,
         });
       }
     };
