+++ src/common/modals/AccountsManager.js
@@ -1,6 +1,6 @@
 import React from "react";
 import styled from "styled-components";
-import { Spin, message } from "antd";
+import { Spin, Button, message } from "antd";
 import { useSelector, useDispatch } from "react-redux";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faTrash } from "@fortawesome/free-solid-svg-icons";
@@ -110,17 +110,26 @@ const ProfileSettings = () => {
                     }
                   `}
                 >
-                  <FontAwesomeIcon
+                  <Button
+                    type="primary"
                     onClick={async () => {
                       const result = await dispatch(
-                        removeAccount(account.selectedProfile.id)
+                        openModal("ActionConfirmation", {
+                          message:
+                            "Are you sure you want to remove this account?",
+                          confirmCallback: removeAccount(
+                            account.selectedProfile.id
+                          ),
+                          title: "Confirm",
+                        })
                       );
                       if (!result) {
                         dispatch(closeModal());
                       }
                     }}
-                    icon={faTrash}
-                  />
+                  >
+                    <FontAwesomeIcon icon={faTrash} />
+                  </Button>
                 </div>
               </AccountContainer>
             );
