+++ src/common/modals/Settings/components/Java.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from "react";
+import React, { useEffect } from "react";
 import { useSelector, useDispatch } from "react-redux";
 import styled from "styled-components";
 import { ipcRenderer } from "electron";
@@ -8,21 +8,15 @@ import {
   faMemory,
   faFolder,
   faUndo,
-  faLevelDownAlt,
   faList,
-  faDesktop,
 } from "@fortawesome/free-solid-svg-icons";
-import { Slider, Button, Input, Switch, Select } from "antd";
+import { Slider, Button, Input, Switch } from "antd";
 import {
   updateJavaArguments,
   updateJavaMemory,
   updateJavaPath,
-  updateResolution,
 } from "../../../reducers/settings/actions";
-import {
-  DEFAULT_JAVA_ARGS,
-  resolutionPresets,
-} from "../../../../app/desktop/utils/constants";
+import { DEFAULT_JAVA_ARGS } from "../../../../app/desktop/utils/constants";
 import { _getJavaPath } from "../../../utils/selectors";
 import { openModal } from "../../../reducers/modals/actions";
 
@@ -42,27 +36,6 @@ const AutodetectPath = styled.div`
 const SelectMemory = styled.div`
   width: 100%;
   height: 100px;
-`;
-
-const Resolution = styled.div`
-  width: 100%;
-  height: 100px;
-`;
-
-const ResolutionInputContainer = styled.div`
-  margin-top: 10px;
-  width: 100%;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-
-  div {
-    width: 200px;
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-  }
 `;
 
 const JavaCustomArguments = styled.div`
@@ -99,28 +72,26 @@ function resetJavaArguments(dispatch) {
 }
 
 const marks = {
-  2048: "2048 MB",
-  4096: "4096 MB",
-  8192: "8192 MB",
-  16384: "16384 MB",
+  2048: "2GiB",
+  4096: "4GiB",
+  6114: "6GiB",
+  8192: "8GiB",
+  10240: "10GiB",
+  12288: "12GiB",
+  14336: "14GiB",
+  16384: "16GiB",
 };
 
 export default function MyAccountPreferences() {
-  const [screenResolution, setScreenResolution] = useState(null);
   const javaArgs = useSelector((state) => state.settings.java.args);
   const javaMemory = useSelector((state) => state.settings.java.memory);
   const javaPath = useSelector(_getJavaPath);
   const customJavaPath = useSelector((state) => state.settings.java.path);
-  const mcResolution = useSelector(
-    (state) => state.settings.minecraftSettings.resolution
-  );
+
   const dispatch = useDispatch();
 
   useEffect(() => {
-    ipcRenderer
-      .invoke("getAllDisplaysBounds")
-      .then(setScreenResolution)
-      .catch(console.error);
+    ipcRenderer.invoke("getAllDisplaysBounds").catch(console.error);
   }, []);
 
   return (
@@ -175,18 +146,14 @@ export default function MyAccountPreferences() {
             <div
               css={`
                 width: 100%;
-                margin: 20px 0;
+                margin: 10px 0px;
               `}
             >
-              <FontAwesomeIcon
-                icon={faLevelDownAlt}
-                flip="horizontal"
-                transform={{ rotate: 90 }}
-              />
               <Input
                 css={`
-                  width: 75%;
-                  margin: 0 10px;
+                  width: 90%;
+                  height: 32px;
+                  float: left;
                 `}
                 onChange={(e) => dispatch(updateJavaPath(e.target.value))}
                 value={customJavaPath}
@@ -208,75 +175,6 @@ export default function MyAccountPreferences() {
           </div>
         </>
       )}
-      <Hr />
-      <Resolution>
-        <Title
-          css={`
-            width: 100%;
-            margin-top: 0px;
-            height: 8px;
-            text-align: left;
-            margin-bottom: 20px;
-          `}
-        >
-          Game Resolution&nbsp; <FontAwesomeIcon icon={faDesktop} />
-        </Title>
-        <Paragraph
-          css={`
-            width: 100%;
-            text-align: left;
-            margin: 0;
-          `}
-        >
-          Select the initial game resolution in pixels (width x height)
-        </Paragraph>
-        <ResolutionInputContainer>
-          <div>
-            <Input
-              placeholder="width"
-              value={mcResolution.width}
-              onChange={(e) => {
-                const w = parseInt(e.target.value, 10);
-                dispatch(updateResolution({ width: w || 854 }));
-              }}
-            />
-            &nbsp;X&nbsp;
-            <Input
-              placeholder="Height"
-              value={mcResolution.height}
-              onChange={(e) => {
-                const h = parseInt(e.target.value, 10);
-                dispatch(updateResolution({ height: h || 480 }));
-              }}
-            />
-          </div>
-          <Select
-            placeholder="Presets"
-            onChange={(v) => {
-              const w = parseInt(v.split("x")[0], 10);
-              const h = parseInt(v.split("x")[1], 10);
-              dispatch(updateResolution({ height: h, width: w }));
-            }}
-          >
-            {resolutionPresets.map((v) => {
-              const w = parseInt(v.split("x")[0], 10);
-              const h = parseInt(v.split("x")[1], 10);
-
-              const isBiggerThanScreen = (screenResolution || []).every(
-                (bounds) => {
-                  return bounds.width < w || bounds.height < h;
-                }
-              );
-              if (isBiggerThanScreen) return null;
-              return (
-                <Select.Option key={v} value={v}>
-                  {v}
-                </Select.Option>
-              );
-            })}
-          </Select>
-        </ResolutionInputContainer>
-      </Resolution>
       <Hr />
       <SelectMemory>
         <Title
@@ -341,7 +239,7 @@ export default function MyAccountPreferences() {
             onChange={(e) => dispatch(updateJavaArguments(e.target.value))}
             value={javaArgs}
             css={`
-              width: 83%;
+              width: 90%;
               height: 32px;
               float: left;
             `}
