+++ src/common/modals/Settings/index.js
@@ -4,6 +4,9 @@ import { Button } from "antd";
 import { useDispatch } from "react-redux";
 import Modal from "../../components/Modal";
 import General from "./components/General";
+import Caching from "./components/Caching";
+import Privacy from "./components/Privacy";
+import Graphics from "./components/Graphics"
 import Java from "./components/Java";
 import CloseButton from "../../components/CloseButton";
 import { closeModal } from "../../reducers/modals/actions";
@@ -85,6 +88,12 @@ function Page(page) {
       return <General />;
     case "Java":
       return <Java />;
+    case "Caching":
+      return <Caching />;
+    case "Privacy":
+      return <Privacy />;
+    case "Graphics":
+      return <Graphics />;
     default:
       return null;
   }
@@ -124,12 +133,25 @@ export default function Settings() {
           >
             Java
           </SettingsButton>
-          {/* <SettingsButton onClick={() => setPage("User Interface")}>
-            User Interface
+          <SettingsButton
+            active={page === "Caching"}
+            onClick={() => setPage("Caching")}
+          >
+            Caching
+          </SettingsButton>
+          <SettingsButton
+            active={page === "Privacy"}
+            onClick={() => setPage("Privacy")}
+          >
+            Privacy
           </SettingsButton>
           <SettingsTitle>Game Settings</SettingsTitle>
-          <SettingsButton>Graphic Settings</SettingsButton>
-          <SettingsButton>Sound Settings</SettingsButton> */}
+          <SettingsButton
+            active={page === "Graphics"}
+            onClick={() => setPage("Graphics")}
+          >
+            Graphics Settings
+          </SettingsButton>
         </SideMenu>
         <SettingsContainer>
           <SettingsColumn>
